refactor(types): re-home coercion family + resolve BinOp inversion (#1062)
Move the coercion family — cast_catalog, operator_catalog,
function_catalog, coercion_spine, and coerce — byte-faithfully into
reddb-io-types (ADR 0052), with their embedded tests.

Resolve the spine's one external dependency (the query AST's BinOp)
via the PRD's preferred path: the operator vocabulary moves into the
types crate (new `operator` module) and `storage::query::ast`
re-exports it, so every `ast::BinOp` call-site stays untouched.
operator_catalog + function_catalog travel along because the spine
imports them and function_catalog is mutually recursive with it.

The server keeps re-export shims at the historical
`storage::schema::*` paths so the ~call-sites stay untouched.
