+++ crates/reddb-server/src/storage/query/ast.rs
@@ -92,46 +92,15 @@ impl Span {
     }
 }
 
-/// Syntactic binary operators. Parsed precedence determines grouping;
-/// this enum only identifies the operator itself. Comparison and logical
-/// operators live alongside arithmetic so a single `Expr::BinaryOp`
-/// walker can cover every infix form the parser emits.
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum BinOp {
-    // Arithmetic
-    Add,
-    Sub,
-    Mul,
-    Div,
-    Mod,
-    // String
-    Concat,
-    // Comparison
-    Eq,
-    Ne,
-    Lt,
-    Le,
-    Gt,
-    Ge,
-    // Logical
-    And,
-    Or,
-}
-
-impl BinOp {
-    /// Left-binding precedence for Pratt parsing. Higher = binds tighter.
-    /// Mirrors PG gram.y's precedence table for the operators we have.
-    pub fn precedence(self) -> u8 {
-        match self {
-            BinOp::Or => 10,
-            BinOp::And => 20,
-            BinOp::Eq | BinOp::Ne | BinOp::Lt | BinOp::Le | BinOp::Gt | BinOp::Ge => 30,
-            BinOp::Concat => 40,
-            BinOp::Add | BinOp::Sub => 50,
-            BinOp::Mul | BinOp::Div | BinOp::Mod => 60,
-        }
-    }
-}
+/// Syntactic binary operators — the operator vocabulary.
+///
+/// `BinOp` was re-homed into the neutral keystone crate
+/// [`reddb_types::operator`] (ADR 0052): the coercion spine that keys
+/// overload resolution on it now lives there, and leaving the enum behind
+/// would force that crate to depend back on `reddb-server`. This re-export
+/// keeps the historical `ast::BinOp` path resolving so every parser /
+/// analyzer / evaluator call-site stays untouched.
+pub use reddb_types::operator::BinOp;
 
 /// Unary operators — only the two real unaries SQL actually has.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
